+++ client/src/views/all-games/allGamesSlice.ts
@@ -1,6 +1,6 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { AllGamesState } from "client/typings/redux.typings";
-import { GameWithUsernames } from "shared/typings/api/games";
+import { GameWithUsernames } from "shared/typings/models/game";
 
 const initialState: AllGamesState = { games: [], signups: [] };
