+++ client/src/views/results/resultsUtils.ts
@@ -1,5 +1,5 @@
 import { UsersForGame } from "client/typings/redux.typings";
-import { UserSignup } from "shared/typings/api/games";
+import { UserSignup } from "shared/typings/models/game";
 
 export const getUsersForGameId = (
   gameId: string,
