+++ shared/typings/api/assignment.ts
@@ -1,6 +1,8 @@
 import { z } from "zod";
 import { Result } from "shared/typings/models/result";
 
+// POST player assignment
+
 export const PostPlayerAssignmentRequestSchema = z.object({
   startingTime: z.string().min(1),
 });
