+++ shared/typings/api/favorite.ts
@@ -1,6 +1,8 @@
 import { z } from "zod";
 import { Game } from "shared/typings/models/game";
 
+// POST favorite
+
 export const PostFavoriteRequestSchema = z.object({
   username: z.string(),
   favoritedGameIds: z.array(z.string()),
