+++ shared/typings/api/games.ts
@@ -1,5 +1,7 @@
 import { ApiError } from "shared/typings/api/errors";
-import { Game } from "shared/typings/models/game";
+import { Game, GameWithUsernames } from "shared/typings/models/game";
+
+// POST update games
 
 export interface PostUpdateGamesResponse {
   message: string;
@@ -11,6 +13,8 @@ export interface PostUpdateGamesError extends ApiError {
   errorId: "unknown";
 }
 
+// GET games
+
 export interface GetGamesResponse {
   message: string;
   status: "success";
@@ -19,14 +23,4 @@ export interface GetGamesResponse {
 
 export interface GetGamesError extends ApiError {
   errorId: "unknown";
-}
-
-export interface GameWithUsernames {
-  game: Game;
-  users: UserSignup[];
-}
-
-export interface UserSignup {
-  username: string;
-  signupMessage: string;
 }
