+++ shared/typings/api/login.ts
@@ -2,6 +2,8 @@ import { z } from "zod";
 import { ApiError } from "shared/typings/api/errors";
 import { UserGroup } from "shared/typings/models/user";
 
+// POST login
+
 export const PostLoginRequestSchema = z.object({
   username: z.string(),
   password: z.string(),
@@ -22,6 +24,8 @@ export interface PostLoginResponse {
 export interface PostLoginError extends ApiError {
   errorId: "unknown" | "loginFailed" | "loginDisabled";
 }
+
+// POST session recovery
 
 export interface PostSessionRecoveryRequest {
   jwt: string;
