+++ shared/typings/api/results.ts
@@ -1,6 +1,8 @@
 import { z } from "zod";
 import { Result } from "shared/typings/models/result";
 
+// GET results
+
 export const GetResultsRequestSchema = z.object({
   startTime: z.string(),
 });
