+++ shared/typings/api/users.ts
@@ -2,6 +2,15 @@ import { z } from "zod";
 import { ApiError } from "shared/typings/api/errors";
 import { SignupMessage } from "shared/typings/models/signupMessage";
 import { UserGames } from "shared/typings/models/user";
+import {
+  PASSWORD_LENGTH_MAX,
+  PASSWORD_LENGTH_MIN,
+  USERNAME_LENGTH_MAX,
+  USERNAME_LENGTH_MIN,
+} from "shared/constants/validation";
+import { sharedConfig } from "shared/config/sharedConfig";
+
+// GET user
 
 export const GetUserRequestSchema = z.object({
   username: z.string(),
@@ -17,11 +26,25 @@ export interface GetUserResponse {
   username: string;
 }
 
-export interface PostUserRequest {
-  username: string;
-  password: string;
-  serial?: string;
-}
+// POST user
+
+export const PostUserRequestSchema = z.object({
+  username: z.string().trim().min(USERNAME_LENGTH_MIN).max(USERNAME_LENGTH_MAX),
+  password: z.string().trim().min(PASSWORD_LENGTH_MIN).max(PASSWORD_LENGTH_MAX),
+  serial: z
+    .string()
+    .optional()
+    .refine((input) => {
+      if (sharedConfig.requireRegistrationCode) {
+        if (!input || input.trim().length === 0) {
+          return false;
+        }
+      }
+      return true;
+    }),
+});
+
+export type PostUserRequest = z.infer<typeof PostUserRequestSchema>;
 
 export interface PostUserResponse {
   message: string;
@@ -34,6 +57,8 @@ export interface PostUserError extends ApiError {
   errorId: "unknown" | "invalidSerial" | "usernameNotFree";
 }
 
+// POST update user password
+
 export const PostUpdateUserPasswordRequestSchema = z.object({
   username: z.string(),
   password: z.string(),
@@ -45,6 +70,8 @@ export type PostUpdateUserPasswordRequest = z.infer<
 >;
 
 export type PostUpdateUserPasswordResponse = PostUserResponse;
+
+// GET user by serial
 
 export const GetUserBySerialRequestSchema = z.object({
   searchTerm: z.string(),
@@ -60,6 +87,8 @@ export interface GetUserBySerialResponse {
   status: "success";
   username: string;
 }
+
+// GET signup messages
 
 export interface GetSignupMessagesResponse {
   signupMessages: SignupMessage[];
